requirements.txt
chainlit==1.2.0
fastapi==0.112.4
fastapi-cli==0.0.5
faiss_cpu==1.8.0.post1
langchain==0.3.1
langchain_anthropic==0.2.1
langchain_chroma==0.1.4
langchain_cohere==0.3.0
langchain_community==0.3.1
langchain_core==0.3.6
langchain_huggingface==0.1.0
langchain_ollama==0.2.0
langchain_openai==0.2.1
pydantic==2.9.2
pypdf==5.0.0 
elevenlabs=1.10.0